Rebalance the 30-gold economy and show summon odds

+++ src/main/java/dev/moma/core/EnemySpawn.java
@@ -2,6 +2,6 @@ package dev.moma.core;
 
 import java.util.UUID;
 
-public record EnemySpawn(EnemyType type, double health, double speed, long reward, boolean boss) {
+public record EnemySpawn(EnemyType type, double health, double speed, double reward, boolean boss) {
     public Enemy create(UUID entity, String arena) { return new Enemy(entity, arena, type, health, speed, reward, boss); }
 }
